# src/infra/swagger/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/health": {
      "get": {
        "operationId": "HealthController_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Servidor e banco de dados disponíveis"
          }
        },
        "tags": [
          "health"
        ]
      }
    },
    "/api/matches": {
      "post": {
        "operationId": "MatchesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit por IP excedido ou TOO_MANY_LOOKUPS."
          }
        },
        "tags": [
          "matches"
        ]
      }
    },
    "/api/matches/{code}": {
      "get": {
        "operationId": "MatchesController_get",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit por IP excedido ou TOO_MANY_LOOKUPS."
          }
        },
        "tags": [
          "matches"
        ]
      }
    },
    "/api/matches/{code}/reshuffle": {
      "post": {
        "operationId": "MatchesController_reshuffle",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit por IP excedido ou TOO_MANY_LOOKUPS."
          }
        },
        "tags": [
          "matches"
        ]
      }
    },
    "/api/matches/{code}/swap": {
      "post": {
        "operationId": "MatchesController_swap",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwapPlayersDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit por IP excedido ou TOO_MANY_LOOKUPS."
          }
        },
        "tags": [
          "matches"
        ]
      }
    },
    "/api/matches/{code}/start": {
      "post": {
        "operationId": "MatchesController_start",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit por IP excedido ou TOO_MANY_LOOKUPS."
          }
        },
        "tags": [
          "matches"
        ]
      }
    },
    "/api/matches/{code}/games/win": {
      "post": {
        "operationId": "MatchesController_win",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoserTeamDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit por IP excedido ou TOO_MANY_LOOKUPS."
          }
        },
        "tags": [
          "matches"
        ]
      }
    },
    "/api/matches/{code}/games/draw": {
      "post": {
        "operationId": "MatchesController_draw",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MatchView"
                    },
                    {
                      "$ref": "#/components/schemas/PenaltiesRequiredView"
                    }
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Rate limit por IP excedido ou TOO_MANY_LOOKUPS."
          }
        },
        "tags": [
          "matches"
        ]
      }
    },
    "/api/matches/{code}/games/penalties": {
      "post": {
        "operationId": "MatchesController_penalties",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoserTeamDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit por IP excedido ou TOO_MANY_LOOKUPS."
          }
        },
        "tags": [
          "matches"
        ]
      }
    },
    "/api/matches/{code}/players": {
      "post": {
        "operationId": "MatchesController_join",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinPlayerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit por IP excedido ou TOO_MANY_LOOKUPS."
          }
        },
        "tags": [
          "matches"
        ]
      }
    },
    "/api/matches/{code}/players/{playerId}": {
      "delete": {
        "operationId": "MatchesController_leave",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "playerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fallback",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "reduce-team-size"
              ]
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit por IP excedido ou TOO_MANY_LOOKUPS."
          }
        },
        "tags": [
          "matches"
        ]
      }
    },
    "/api/matches/{code}/team-size": {
      "put": {
        "operationId": "MatchesController_changeTeamSize",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeTeamSizeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit por IP excedido ou TOO_MANY_LOOKUPS."
          }
        },
        "tags": [
          "matches"
        ]
      }
    },
    "/api/matches/{code}/end": {
      "post": {
        "operationId": "MatchesController_end",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit por IP excedido ou TOO_MANY_LOOKUPS."
          }
        },
        "tags": [
          "matches"
        ]
      }
    },
    "/api/matches/{code}/timer/start": {
      "post": {
        "operationId": "MatchesController_startTimer",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit por IP excedido ou TOO_MANY_LOOKUPS."
          }
        },
        "tags": [
          "matches"
        ]
      }
    },
    "/api/matches/{code}/timer/pause": {
      "post": {
        "operationId": "MatchesController_pauseTimer",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit por IP excedido ou TOO_MANY_LOOKUPS."
          }
        },
        "tags": [
          "matches"
        ]
      }
    },
    "/api/matches/{code}/timer/reset": {
      "post": {
        "operationId": "MatchesController_resetTimer",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit por IP excedido ou TOO_MANY_LOOKUPS."
          }
        },
        "tags": [
          "matches"
        ]
      }
    },
    "/api/matches/{code}/undo": {
      "post": {
        "operationId": "MatchesController_undo",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit por IP excedido ou TOO_MANY_LOOKUPS."
          }
        },
        "tags": [
          "matches"
        ]
      }
    }
  },
  "info": {
    "title": "Clanada API",
    "description": "API de gerenciamento de pelada com rotação de quadras",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "MatchConfigDto": {
        "type": "object",
        "properties": {
          "teamSize": {
            "type": "number",
            "minimum": 2
          },
          "colors": {
            "type": "array",
            "minItems": 2,
            "items": {
              "type": "string",
              "enum": [
                "verde",
                "vermelho",
                "azul",
                "amarelo",
                "laranja",
                "preto",
                "branco",
                "rosa",
                "roxo",
                "cinza"
              ]
            }
          },
          "gameMinutes": {
            "type": "number",
            "minimum": 1,
            "maximum": 60
          }
        },
        "required": [
          "teamSize",
          "colors",
          "gameMinutes"
        ]
      },
      "CreateMatchDto": {
        "type": "object",
        "properties": {
          "config": {
            "$ref": "#/components/schemas/MatchConfigDto"
          },
          "playerNames": {
            "minItems": 1,
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "config",
          "playerNames"
        ]
      },
      "PlayerView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "TeamView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "players": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayerView"
            }
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "gameStreak": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "players",
          "color",
          "gameStreak"
        ]
      },
      "TimerView": {
        "type": "object",
        "properties": {
          "startedAt": {
            "type": "string",
            "nullable": true
          },
          "elapsedMs": {
            "type": "number"
          }
        },
        "required": [
          "startedAt",
          "elapsedMs"
        ]
      },
      "MatchView": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "ENDED"
            ]
          },
          "version": {
            "type": "number"
          },
          "config": {
            "$ref": "#/components/schemas/MatchConfigDto"
          },
          "teams": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamView"
            }
          },
          "queue": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayerView"
            }
          },
          "timer": {
            "$ref": "#/components/schemas/TimerView"
          },
          "serverNow": {
            "type": "string"
          },
          "canUndo": {
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "status",
          "version",
          "config",
          "teams",
          "queue",
          "timer",
          "serverNow",
          "canUndo"
        ]
      },
      "SwapPlayersDto": {
        "type": "object",
        "properties": {
          "playerAId": {
            "type": "string"
          },
          "playerBId": {
            "type": "string"
          }
        },
        "required": [
          "playerAId",
          "playerBId"
        ]
      },
      "LoserTeamDto": {
        "type": "object",
        "properties": {
          "loserTeamId": {
            "type": "string"
          }
        },
        "required": [
          "loserTeamId"
        ]
      },
      "PenaltiesRequiredView": {
        "type": "object",
        "properties": {
          "penaltiesRequired": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "penaltiesRequired"
        ]
      },
      "JoinPlayerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "ChangeTeamSizeDto": {
        "type": "object",
        "properties": {
          "teamSize": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "teamSize"
        ]
      }
    }
  }
}